[BE] refactor: findAllByAdmin 삭제 기본 findAll(Pageable) 사용 #96
- JpaRepository에 이미 Pageable을 파라미터로 받는 findAll()이 정의되어 있으므로 불필요한 코드임
