+++ backend/src/main/java/com/onair/hearit/admin/application/AdminHearitService.java
@@ -37,7 +37,7 @@ public class AdminHearitService {
     //TODO: 리팩터링 필요
     public PagedResponse<HearitAdminResponse> getPageHearits(int page, int size) {
         Pageable pageable = PageRequest.of(page, size, Sort.by(Sort.Order.asc("createdAt")));
-        Page<Hearit> pageHearits = hearitRepository.findAllForAdmin(pageable);
+        Page<Hearit> pageHearits = hearitRepository.findAll(pageable);
         List<Long> hearitIds = pageHearits.getContent().stream()
                 .map(Hearit::getId)
                 .toList();
