+++ backend/src/main/java/com/onair/hearit/admin/presentation/AdminCategoryController.java
@@ -3,7 +3,7 @@ package com.onair.hearit.admin.presentation;
 import com.onair.hearit.admin.application.AdminCategoryService;
 import com.onair.hearit.admin.dto.request.CategoryCreateRequest;
 import com.onair.hearit.admin.dto.response.PagedResponse;
-import com.onair.hearit.dto.response.CategoryInfoResponse;
+import com.onair.hearit.admin.dto.response.CategoryInfoResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
